refactor: DeviceSearchPage.tsx 매핑 함수, 로직, 모달 UI 분리
